requirements.txt
rdflib
SPARQLWrapper
requests
python-dotenv